Extract lock sugar samples
